feat: enhance product retrieval and search functionality; implement advanced filtering, pagination, and search in getProducts, update product model for text indexing, and improve search modal with debounced API calls

+++ frontend/src/components/header/CartItem.jsx
@@ -9,9 +9,12 @@ const CartItem = ({ currProd, products }) => {
   const [quantity, setQuantity] = useState(currProd.quantity);
 
   const { serverURL } = useAuth();
-  const { setCart } = useCart();
+  const { setCart, loading, setLoading } = useCart();
 
   const handleQuantityMinus = async () => {
+    if (loading) return;
+    setLoading(true);
+
     if (quantity <= 1) return;
     setQuantity(quantity - 1);
 
@@ -23,7 +26,7 @@ const CartItem = ({ currProd, products }) => {
           productId: currProd.product._id,
           variant: currProd.variant,
         },
-        { withCredentials: true }
+        { withCredentials: true },
       );
 
       setCart(data);
@@ -31,10 +34,15 @@ const CartItem = ({ currProd, products }) => {
       console.error("Cart Error:", error?.response?.data || error.message);
       setQuantity(quantity);
       notifyError("Quantity update failed");
+    } finally {
+      setLoading(false);
     }
   };
 
   const handleQuantityPlus = async () => {
+    if (loading) return;
+    setLoading(true);
+
     if (quantity >= 10) return;
     setQuantity(quantity + 1);
 
@@ -46,7 +54,7 @@ const CartItem = ({ currProd, products }) => {
           productId: currProd.product._id,
           variant: currProd.variant,
         },
-        { withCredentials: true }
+        { withCredentials: true },
       );
 
       setCart(data);
@@ -54,10 +62,14 @@ const CartItem = ({ currProd, products }) => {
       console.log("Cart Error:", error?.response?.data || error.message);
       setQuantity(quantity);
       notifyError("Quantity update failed");
+    } finally {
+      setLoading(false);
     }
   };
 
   const handleDeleteCart = async () => {
+    if (loading) return;
+    setLoading(true);
     try {
       const { data } = await axios.delete(`${serverURL}/api/cart`, {
         data: {
@@ -73,6 +85,8 @@ const CartItem = ({ currProd, products }) => {
       console.log("Delete Cart Error:", error?.response?.data || error.message);
       notifyError("Cart remove failed");
       setQuantity(quantity);
+    } finally {
+      setLoading(false);
     }
   };
